server1/requirements.txt
-i https://pypi.org/simple

cryptography==42.0.8; python_version >= '3.7'
django==4.2.14; python_version >= '3.8'
pycryptodome==3.20.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
requests[socks]==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
